invoker: wf-1779680045111-2/implement-ui-recreate-downstream-action — Add the task context menu action and UI tests for Recreate Downstream. Layer: ui_activation Feature state: active Goal: Let users trigger downstream-only recreate from the task context menu. Motivation: The requested workflow is naturally task-local: right-click the task whose branch was manually updated, then rerun its subsequent tasks. Alternative considerations: A workflow context menu item was rejected because the operation is rooted at one task's downstream subgraph. Implementation details: Add "Recreate Downstream" beside "Recreate from Task" in the task danger section, wire it to window.invoker.recreateDownstream, and update mocks/types/tests. Review claim: The UI exposes a clear task-level action for downstream-only recreation without changing existing task or workflow menu behavior. Safety invariant: Existing Restart Task and Recreate from Task actions retain their current labels, ordering, handlers, and semantics. Slice rationale: UI activation depends on the app bridge contract and can be reviewed separately from core mutation semantics. Architectural effect: Renderer task context actions gain one new command path through the generated Invoker API. Files: - packages/ui/src/lib/context-menu-items.ts - packages/ui/src/components/ContextMenu.tsx - packages/ui/src/App.tsx - packages/ui/src/__tests__/helpers/mock-invoker.ts - packages/ui/src/__tests__/context-menu.test.ts - packages/ui/src/__tests__/context-menu-e2e.test.tsx Change types: - packages/ui/src/lib/context-menu-items.ts: modify - packages/ui/src/components/ContextMenu.tsx: modify - packages/ui/src/App.tsx: modify - packages/ui/src/__tests__/helpers/mock-invoker.ts: test-only - packages/ui/src/__tests__/context-menu.test.ts: test-only - packages/ui/src/__tests__/context-menu-e2e.test.tsx: test-only Acceptance criteria: - Task context menus show "Recreate Downstream" for workflow-owned non-running tasks. - Clicking the menu item calls `window.invoker.recreateDownstream(task.id)`. - Workflow context menus do not show the task-scoped action. - Focused UI tests exit code 0.
Context:
  wf-1779680045111-2/implement-app-bridge-recreate-downstream (3ab657a): Expose recreateDownstream through contracts, main IPC, mutation facade, API/headless routing, and focused app tests.
Layer: app_bridge
Feature state: active
Goal: Make the new domain mutation available to all non-renderer surfaces that can trigger task mutations.
Motivation: The UI action and CLI command need one shared app bridge so dispatch/topup, serialization, and owner-process behavior stay consistent.
Alternative considerations: UI-only IPC was rejected because users also asked to submit scriptable Invoker work, and headless recovery is useful for the same operational case.
Implementation details: Add an IPC channel and headless command for recreate-downstream, route it through the same mutation lifecycle as recreate-task, and document help text.
Review claim: app surfaces can invoke recreateDownstream without bypassing existing mutation serialization, preemption, dispatch, or topup paths.
Safety invariant: Existing recreate-task and retry-task channels continue to call their existing primitives unchanged.
Slice rationale: This depends on the core primitive and provides a stable contract before UI activation.
Architectural effect: The app bridge gains a new task mutation command that flows from contracts/preload/headless/main into workflow-core via the shared mutation facade.
Files:
- packages/contracts/src/ipc-channels.ts
- packages/app/src/workflow-actions.ts
- packages/app/src/workflow-mutation-facade.ts
- packages/app/src/main.ts
- packages/app/src/headless.ts
- packages/app/src/__tests__/headless-client.test.ts
- packages/app/src/__tests__/api-server.test.ts
Change types:
- packages/contracts/src/ipc-channels.ts: modify
- packages/app/src/workflow-actions.ts: modify
- packages/app/src/workflow-mutation-facade.ts: modify
- packages/app/src/main.ts: modify
- packages/app/src/headless.ts: modify
- packages/app/src/__tests__/headless-client.test.ts: test-only
- packages/app/src/__tests__/api-server.test.ts: test-only
Acceptance criteria:
- `window.invoker.recreateDownstream` is generated from the IPC registry.
- `./run.sh --headless recreate-downstream <taskId>` is accepted and supports `--no-track`.
- Main IPC dispatches ready downstream tasks and runs global topup through the existing facade lifecycle.
- Focused app tests exit code 0.

Solution:
  Add the task context menu action and UI tests for Recreate Downstream.
Layer: ui_activation
Feature state: active
Goal: Let users trigger downstream-only recreate from the task context menu.
Motivation: The requested workflow is naturally task-local: right-click the task whose branch was manually updated, then rerun its subsequent tasks.
Alternative considerations: A workflow context menu item was rejected because the operation is rooted at one task's downstream subgraph.
Implementation details: Add "Recreate Downstream" beside "Recreate from Task" in the task danger section, wire it to window.invoker.recreateDownstream, and update mocks/types/tests.
Review claim: The UI exposes a clear task-level action for downstream-only recreation without changing existing task or workflow menu behavior.
Safety invariant: Existing Restart Task and Recreate from Task actions retain their current labels, ordering, handlers, and semantics.
Slice rationale: UI activation depends on the app bridge contract and can be reviewed separately from core mutation semantics.
Architectural effect: Renderer task context actions gain one new command path through the generated Invoker API.
Files:
- packages/ui/src/lib/context-menu-items.ts
- packages/ui/src/components/ContextMenu.tsx
- packages/ui/src/App.tsx
- packages/ui/src/__tests__/helpers/mock-invoker.ts
- packages/ui/src/__tests__/context-menu.test.ts
- packages/ui/src/__tests__/context-menu-e2e.test.tsx
Change types:
- packages/ui/src/lib/context-menu-items.ts: modify
- packages/ui/src/components/ContextMenu.tsx: modify
- packages/ui/src/App.tsx: modify
- packages/ui/src/__tests__/helpers/mock-invoker.ts: test-only
- packages/ui/src/__tests__/context-menu.test.ts: test-only
- packages/ui/src/__tests__/context-menu-e2e.test.tsx: test-only
Acceptance criteria:
- Task context menus show "Recreate Downstream" for workflow-owned non-running tasks.
- Clicking the menu item calls `window.invoker.recreateDownstream(task.id)`.
- Workflow context menus do not show the task-scoped action.
- Focused UI tests exit code 0.
