+++ packages/ui/src/App.tsx
@@ -9,7 +9,7 @@
  * - Modals overlay when needed
  */
 
-import { useState, useCallback, useMemo, useEffect, useRef, useLayoutEffect } from 'react';
+import { useState, useCallback, useMemo, useEffect, useRef } from 'react';
 import yaml from 'js-yaml';
 import type { TaskState, TaskReplacementDef, ExternalGatePolicyUpdate, WorkflowMeta, WorkflowStatus } from './types.js';
 import type { ActionGraphNode, TerminalSessionDescriptor } from '@invoker/contracts';
@@ -22,6 +22,7 @@ import { ApprovalModal } from './components/ApprovalModal.js';
 import { InputModal } from './components/InputModal.js';
 import { ExperimentModal } from './components/ExperimentModal.js';
 import { ContextMenu } from './components/ContextMenu.js';
+import { WorkflowContextMenu } from './components/WorkflowContextMenu.js';
 import { QueueView } from './components/QueueView.js';
 import { ReplaceTaskModal } from './components/ReplaceTaskModal.js';
 import { SystemSetupModal } from './components/SystemSetupModal.js';
@@ -80,153 +81,6 @@ function isEditableKeyboardTarget(target: EventTarget | null): boolean {
 
 function normalizedSearchText(value: string | undefined): string {
   return (value ?? '').toLowerCase();
-}
-
-interface WorkflowContextMenuProps {
-  x: number;
-  y: number;
-  workflowId: string;
-  onOpenWorkflow: (workflowId: string) => void;
-  onOpenPr: (workflowId: string) => void;
-  onRetryWorkflow: (workflowId: string) => void;
-  onRebaseRetry: (workflowId: string) => void;
-  onRebaseRecreate: (workflowId: string) => void;
-  onRecreateWorkflow: (workflowId: string) => void;
-  onCancelWorkflow: (workflowId: string) => void;
-  onDeleteWorkflow: (workflowId: string) => void;
-  onCopyWorkflowId: (workflowId: string) => void;
-  onClose: () => void;
-}
-
-function WorkflowContextMenu({
-  x,
-  y,
-  workflowId,
-  onOpenWorkflow,
-  onOpenPr,
-  onRetryWorkflow,
-  onRebaseRetry,
-  onRebaseRecreate,
-  onRecreateWorkflow,
-  onCancelWorkflow,
-  onDeleteWorkflow,
-  onCopyWorkflowId,
-  onClose,
-}: WorkflowContextMenuProps): JSX.Element {
-  const menuRef = useRef<HTMLDivElement>(null);
-  const [position, setPosition] = useState({ left: x, top: y });
-  const [showMore, setShowMore] = useState(false);
-
-  useLayoutEffect(() => {
-    if (!menuRef.current) return;
-
-    const rect = menuRef.current.getBoundingClientRect();
-    const viewportWidth = window.innerWidth;
-    const viewportHeight = window.innerHeight;
-
-    let left = x;
-    let top = y;
-
-    if (rect.right > viewportWidth) {
-      left = x - rect.width;
-    }
-    if (rect.bottom > viewportHeight) {
-      top = y - rect.height;
-    }
-
-    left = Math.max(0, Math.min(left, viewportWidth - rect.width));
-    top = Math.max(0, Math.min(top, viewportHeight - rect.height));
-    setPosition({ left, top });
-  }, [x, y, showMore]);
-
-  useEffect(() => {
-    const dismissFromOutsideTarget = (target: EventTarget | null, button?: number) => {
-      if (button !== undefined && button !== 0) return;
-      if (menuRef.current && !menuRef.current.contains(target as Node)) {
-        onClose();
-      }
-    };
-    const handlePointerDownCapture = (event: PointerEvent) => dismissFromOutsideTarget(event.target, event.button);
-    const handleMouseDownCapture = (event: MouseEvent) => dismissFromOutsideTarget(event.target, event.button);
-    const handleClickCapture = (event: MouseEvent) => dismissFromOutsideTarget(event.target, event.button);
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') onClose();
-    };
-
-    document.addEventListener('pointerdown', handlePointerDownCapture, true);
-    document.addEventListener('mousedown', handleMouseDownCapture, true);
-    document.addEventListener('click', handleClickCapture, true);
-    document.addEventListener('keydown', handleKeyDown);
-    return () => {
-      document.removeEventListener('pointerdown', handlePointerDownCapture, true);
-      document.removeEventListener('mousedown', handleMouseDownCapture, true);
-      document.removeEventListener('click', handleClickCapture, true);
-      document.removeEventListener('keydown', handleKeyDown);
-    };
-  }, [onClose]);
-
-  const runAction = (action: (workflowId: string) => void) => {
-    action(workflowId);
-    onClose();
-  };
-
-  const buttonClass = 'w-full px-3 py-1.5 text-left text-sm text-gray-100 hover:bg-gray-700';
-  const dangerButtonClass = 'w-full px-3 py-1.5 text-left text-sm text-red-300 hover:bg-gray-700';
-
-  return (
-    <div
-      ref={menuRef}
-      role="menu"
-      className="fixed z-50 min-w-[200px] rounded-lg border border-gray-600 bg-gray-800 py-1 shadow-xl"
-      style={{ left: position.left, top: position.top }}
-      tabIndex={-1}
-      onClick={(event) => event.stopPropagation()}
-    >
-      <button role="menuitem" onClick={() => runAction(onOpenWorkflow)} className={buttonClass}>
-        Open Workflow
-      </button>
-      <button role="menuitem" onClick={() => runAction(onOpenPr)} className={buttonClass}>
-        Open PR
-      </button>
-      <button role="menuitem" onClick={() => runAction(onRetryWorkflow)} className={buttonClass}>
-        Retry Workflow
-      </button>
-      <button role="menuitem" onClick={() => runAction(onCopyWorkflowId)} className={buttonClass}>
-        Copy Workflow ID
-      </button>
-      {!showMore ? (
-        <div>
-          <div className="my-1 border-t border-gray-600" />
-          <button
-            role="menuitem"
-            className="w-full px-3 py-1.5 text-left text-sm text-gray-300 hover:bg-gray-700"
-            onClick={() => setShowMore(true)}
-          >
-            More
-          </button>
-        </div>
-      ) : (
-        <div>
-          <div className="my-1 border-t border-gray-600" />
-          <button role="menuitem" onClick={() => runAction(onRebaseRetry)} className={buttonClass}>
-            Rebase and Retry
-          </button>
-          <button role="menuitem" onClick={() => runAction(onRebaseRecreate)} className={dangerButtonClass}>
-            Rebase and Recreate
-          </button>
-          <button role="menuitem" onClick={() => runAction(onRecreateWorkflow)} className={dangerButtonClass}>
-            Recreate Workflow
-          </button>
-          <button role="menuitem" onClick={() => runAction(onCancelWorkflow)} className={dangerButtonClass}>
-            Cancel Workflow
-          </button>
-          <button role="menuitem" onClick={() => runAction(onDeleteWorkflow)} className={dangerButtonClass}>
-            Delete Workflow
-          </button>
-        </div>
-      )}
-    </div>
-  );
 }
 
 function GearIcon(): JSX.Element {
@@ -1059,6 +913,15 @@ export function App() {
     }
   }, []);
 
+  const handleRecreateDownstream = useCallback(async (taskId: string) => {
+    setContextMenu(null);
+    try {
+      await window.invoker?.recreateDownstream(taskId);
+    } catch (err) {
+      console.error('Recreate Downstream failed:', err);
+    }
+  }, []);
+
   const handleDeleteWorkflow = useCallback(async (workflowId: string) => {
     setContextMenu(null);
     const confirmed = window.confirm(
@@ -1837,6 +1700,7 @@ export function App() {
           onReplace={handleReplaceTask}
           onOpenTerminal={handleOpenTerminal}
           onRecreateTask={handleRecreateTask}
+          onRecreateDownstream={handleRecreateDownstream}
           onFix={handleFix}
           onCancel={handleCancelTask}
           onClose={closeContextMenu}
