+++ packages/ui/src/lib/context-menu-items.ts
@@ -40,6 +40,7 @@ export function getMenuItems(
   const canOpenTerminal = !isExperimentSpawnPivotTask(task);
   const canFix = task.status === 'failed';
   const canCancel = task.status !== 'completed' && task.status !== 'stale';
+  const canRecreateDownstream = Boolean(task.config.workflowId) && task.status !== 'running';
 
   // ── Fix actions (top, only for failed) ────────────────────────
   if (canFix) {
@@ -129,6 +130,15 @@ export function getMenuItems(
       separator: !canCancel ? 'danger' : undefined,
     });
 
+    if (canRecreateDownstream) {
+      items.push({
+        id: 'recreate-downstream',
+        label: 'Recreate Downstream',
+        enabled: true,
+        action: 'onRecreateDownstream',
+        variant: 'danger',
+      });
+    }
   }
 
   return items;
